requirements.txt
# Core
python-dotenv==1.0.0
loguru==0.7.2
pyyaml==6.0.1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# HTTP & Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
google-search-results==2.4.2

# Image Processing
Pillow==10.1.0
imagehash==4.3.1

# Utilities
tqdm==4.66.1

# Validation
pydantic==2.5.3
